requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==4.1.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
numpy==1.25.0
pandas==2.0.3
Pillow==9.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.5.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
